[wip reset this] save google_string on Need
The idea here is to save the original place string a user enters.
This allows it to be re-used for geocoding via the edit workflow.
We got caught because we don't know how to approve our need record in
development.

WARNING:  We have not yet tested this to determine that it is safe
with regard to existing data.

Potential problem workflow:

* open edit page for existing Need with address and lat/lon present
* POTENTIAL PROBLEM HERE: google_place is empty after migration so field
is empty; this may trigger a geocoding on empty value.
* We simply didn't have time to test this workflow.