Administration screens for flagged markers
Screens to allow admins to manage flagged markers.

The app store requires a way to flag any user generated content as inappropriate.
Users can flag any markers on their device as inappropriate. Flagged markers
receive a `inappropriate_flag` key and are hidden from public view. Under the
`inappropriate_flag` key we also capture when it was flagged, the device that
flagged it, and the reason for doing so.

These records will show up on the "Flagged Markers" page where an admin can
either "confirm" or "clear" the marker. An admin will "confirm" the marker if
they feel the content is inappropriate. This sets the `resolved` flag to true
which keeps the marker out of the public view and removes it from the admin
screens.

If an admin finds nothing wrong with the marker, they can "clear" the marker.
This removes the `inappropriate_flag` key and allows the marker to be viewed by
the public once again. Cleared markers are also removed from the admin screen.